cloud_functions/predictor/requirements.txt
firebase-admin>=2.15.1,<2.16.0
google-api-python-client>=1.7.8,<1.8.0
jpholiday>=0.0.6,<0.1.0
